tests/wpt/web-platform-tests/webxr/webGLCanvasContext_create_xrcompatible.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <script src="resources/webxr_test_constants.js"></script>
  <canvas id="webgl-canvas"></canvas>
  <script>
    xr_promise_test("Creating a webglCanvasContext with no device",
      (t) => {
        let webglCanvas = document.createElement('canvas');
        let gl = webglCanvas.getContext('webgl');

        assert_false(gl.getContextAttributes().xrCompatible);
        return promise_rejects_dom(t, "InvalidStateError", gl.makeXRCompatible());
      });

    xr_promise_test("An XR-compatible webglCanvasContext can be created",
      (t) => {
        let gl = null;
        let offscreenGl = null;
        return navigator.xr.test.simulateDeviceConnection(TRACKED_IMMERSIVE_DEVICE)
          .then( (controller) => {
            let webglCanvas = document.getElementById('webgl-canvas');
            gl = webglCanvas.getContext('webgl');
            return gl.makeXRCompatible();
          }).then( () => {
            assert_true(gl.getContextAttributes().xrCompatible);

            // Check that an offscreen context behaves no different.
            let offscreenCanvas = document.createElement('canvas');
            offscreenGl = offscreenCanvas.getContext('webgl');
            return offscreenGl.makeXRCompatible();
          }).then( () => {
            assert_true(offscreenGl.getContextAttributes().xrCompatible);
          });
      });
  </script>
</body>
